refactor(controlplane): move presentation types to api package
server/types.go presentation types (Agent, AgentRuntime, Instance, ... plus the
nested AgentCertificateRecoveryGrantResponse) move to a new leaf package
internal/controlplane/api. server keeps type aliases so its call sites are
unchanged; the storage<->view converters stay in server (api must not import
storage). This breaks the agents<->server import cycle that forced LiveStore to
be generic (collapsed in the next commit).

No behavior change.
